+++ packages/async/src/deferred/deferred.js
@@ -13,13 +13,48 @@
  * Deferred instance
  * @property {function} reject Function to reject the internal promise of the
  * Deferred instance
+ * @property {boolean} settled Flag indicating whether the promise has settled
+ * yet (resolved or rejected)
  */
 class Deferred {
     constructor() {
+        this.settled = false;
         this.promise = new Promise((resolve, reject) => {
-            this.resolve = resolve;
-            this.reject = reject;
+            this.reject = (value) => {
+                this.settled = true;
+                reject(value);
+            };
+            this.resolve = (value) => {
+                this.settled = true;
+                resolve(value);
+            };
         });
+    }
+
+    /**
+     * The `Deferred` equivalent of `Promise.resolve()`. Creates a new `Deferred` instance and instantly
+     * resovles it, then returns the created instance.
+     * 
+     * @param {*} value The value to return as the deferred resolution
+     * @returns {Deferred} Returns the new resolved Deferred instance
+     */
+    static resolve(value) {
+        const deferred = new Deferred();
+        deferred.resolve(value);
+        return deferred;
+    }
+
+    /**
+     * The `Deferred` equivalent of `Promise.reject()`. Creates a new `Deferred` instance and instantly
+     * rejects it, then returns the created instance.
+     * 
+     * @param {*} value The value to return as the deferred rejection
+     * @returns {Deferred} Returns the new rejected Deferred instance
+     */
+    static reject(value) {
+        const deferred = new Deferred();
+        deferred.reject(value);
+        return deferred;
     }
 }